refactor: move some functions from gdpr module to more fitting modules
