+++ .env.example
@@ -1,0 +1,56 @@
+# =======================================
+# SERVER CONFIGURATION
+# =======================================
+APP_NAME=SmartSave
+APP_VERSION=1.0.0
+APP_ENV=... # options: development | production
+
+# =======================================
+# SECURITY & COMPLIANCE
+# =======================================
+JWT_SECRET_KEY=...
+JWT_EXPIRATION_TIME=...# in seconds
+JWT_SIGNING_ALGORITHM=...
+
+MAX_FAILED_LOGIN_ATTEMPTS=...
+IP_HASH_SALT=...
+
+LOG_RETENTION_DAYS=...
+HARD_DELETE_RETENTION_DAYS=...
+HARD_DELETE_CRON_INTERVAL_HOURS=...
+
+MIN_BALANCE_THRESHOLD=...
+MIN_WALLET_DEPOSIT_AMOUNT=...
+MIN_WALLET_WITHDRAWAL_AMOUNT=...
+
+# =======================================
+# DATABASE
+# =======================================
+DB_HOST=...
+DB_PORT=...
+POSTGRES_USER=username
+POSTGRES_PASSWORD=password
+POSTGRES_DB=db-name
+
+# =======================================
+# OAUTH
+# =======================================
+GOOGLE_CLIENT_ID=...
+GOOGLE_CLIENT_SECRET=...
+
+# =======================================
+# EMAIL SETTINGS
+# =======================================
+SMTP_HOST=...
+SMTP_PORT=...
+SMTP_USERNAME=no-reply@example.com
+SMTP_PASSWORD=...
+
+# =======================================
+# OTHER SETTINGS
+# =======================================
+ALLOWED_ORIGINS=...
+FRONTEND_URL=...
+DOCS_USERNAME=...
+DOCS_PASSWORD=...
+TEST_EMAIL_ACCOUNTS=...
